LayoutTests/editing/async-clipboard/clipboard-read-text-from-platform-expected.txt

This test verifies that navigator.clipboard.readText causes DOM paste UI to appear, if the text was written using platform API. To run the test manually, copy text from elsewhere on the system, and then click 'Paste'; DOM paste UI should appear when clicking 'Paste', and pasting should allow the page to read the text that was copied.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Read text from clipboard using navigator.clipboard.readText.
PASS clipboardText is "This text was written by the platform."
PASS successfullyParsed is true

TEST COMPLETE
